+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Integrated Alembic for managing chat history migrations
 - Tool implementations without langchain or langgraph dependencies
 - CRUDs.
+- BlueNaas CRUD tools
 
 ## [0.3.3] - 30.10.2024